refactor: move repository interfaces to core folder
